+++ port_ocean/clients/port/mixins/blueprints.py
@@ -14,13 +14,15 @@ class BlueprintClientMixin:
         self.auth = auth
         self.client = client
 
-    async def get_blueprint(self, identifier: str) -> Blueprint:
+    async def get_blueprint(
+        self, identifier: str, should_log: bool = True
+    ) -> Blueprint:
         logger.info(f"Fetching blueprint with id: {identifier}")
         response = await self.client.get(
             f"{self.auth.api_url}/blueprints/{identifier}",
             headers=await self.auth.headers(),
         )
-        handle_status_code(response)
+        handle_status_code(response, should_log=should_log)
         return Blueprint.parse_obj(response.json()["blueprint"])
 
     async def create_blueprint(
@@ -105,3 +107,30 @@ class BlueprintClientMixin:
         )
 
         handle_status_code(response)
+
+    async def create_page(
+        self,
+        page: dict[str, Any],
+    ) -> dict[str, Any]:
+        logger.info(f"Creating page: {page}")
+        response = await self.client.post(
+            f"{self.auth.api_url}/pages",
+            json=page,
+            headers=await self.auth.headers(),
+        )
+
+        handle_status_code(response)
+        return page
+
+    async def delete_page(
+        self,
+        identifier: str,
+        should_raise: bool = False,
+    ) -> None:
+        logger.info(f"Deleting page: {identifier}")
+        response = await self.client.delete(
+            f"{self.auth.api_url}/pages/{identifier}",
+            headers=await self.auth.headers(),
+        )
+
+        handle_status_code(response, should_raise)