+++ integrations/gitlab/CHANGELOG.md
@@ -7,12 +7,37 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 <!-- towncrier release notes start -->
 
+0.1.41 (2024-01-01)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.4.13 (#1)
+
+### Bug Fixes
+
+- Fixed handling not found error when trying to get tree for commit sha of newly created project (#1)
+
+
+v0.1.40 (2023-12-27)
+====================
+
+### Improvements
+
+- Increased log verbosity for gitlab webhook bootstrap (#2)
+- Added handling for fetching errors (PORT-5813)
+
+### Bug Fixes
+
+- Fixed 404 error when trying to get tree for commit sha of newly created project (#1)
+
+
 v0.1.39 (2023-12-25)
 ====================
 
 ### Improvements
 
-- Fix stale relation identifiers in default blueprints (port-5799)
+- Fix stale relation identifiers in default blueprints (PORT-5799)
 
 
 v0.1.38 (2023-12-24)